Use empty data frames for default slots in SeqVarData
When no variantData is specfied for SeqVarData objects, which is
common, duplicating the variant.id takes unneccesary time and
memory. Instead, create a data frame with the correct number of
rows but 0 columns, and only check for matching variant.id if
required. For consistency, make sampleData behave the same way.