store indices of variant filters in iterator object
Setting a filter using intersections can be computationally
expensive, as can reading chromosome and position on each iteration
to set a filter with ranges. Also, using push/pop for keeping track
of filters prevents setting sample filters on an iterator object.
Change the implementation of the iterator classes to identify the
selection index for each block/range on object creation, and store
a list of indices in the object. This has the added benefit of
allowing all classes to have a single iterator class as parent, and
define only one iterateFilter method on the parent class.