reduce memory usage when creating a SeqVarData object
When validObject() is called during creation of a SeqVarData object,
it calls the sampleData() and variantData() methods, which in turn
call seqGetFilter(), which is then used to index the sample and
variant data.frames. Even for 0-column data.frames, this indexing
creates a copy of a vector that can be very large. Avoid making
this copy if there is no sample or variant selection in place.