correct for seqAlleleCount returning NA when no samples selected
seqAlleleCount is called separately for males and females for
sex chromosomes. Previously it would return 0 when no samples
were selected, but as of SeqArray 1.37.1 it returns NA in that
case. Thus the SeqVarData allele count functions were returning
NA in cases where all samples were male or female. Correct for
that by setting count=NA to 0 when selecting samples by sex.